pyproject.toml
[tool.poetry]
name = "litsearch-backend"
version = "0.1.0"
description = ""
authors = ["Faizan Akhter <faizan.ac@yahoo.com>"]
readme = "README.md"
packages = [{include = "litsearch_backend"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.105.0"}
motor = "^3.3.2"
python-dotenv = "^1.0.0"
uvicorn = "^0.24.0.post1"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
